Reverts "more files to ignore" (commit 1accfdf40f2edf70a70059a61c0a0c8b62d7b3c9).
RestartTest-6 runs the same input files twice:
1. the full length of the beam line
2. restart from the h5-file that was written in the first run and run to
the end. Compare results of run 1 and 2.
This is an ugly hack to test whether the OPAL-T can restart from its
h5-file and yields the same result. This can't be done in a clean way with
the current test infrastructure. The reference is always results from a
prior run.